% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLabelFile_Source.R
\name{classifyQuestions}
\alias{classifyQuestions}
\title{Iterates over list of questions along with question column indices and classifies each question}
\usage{
classifyQuestions(questions, answers, qInds)
}
\arguments{
\item{questions:}{char vector containing all questions column titles}

\item{qInds:}{Beginning and end index of each question column

Notes: The order in which these logical helper functions are called is important to the functionality of this program}
}
\description{
Iterates over list of questions along with question column indices and classifies each question
}
