% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals_Source.R
\name{collapseMcq}
\alias{collapseMcq}
\title{This function takes an MCQ question represented by multiple logical columns and collapses it to a single column}
\usage{
collapseMcq(mcq)
}
\arguments{
\item{mcq:}{question object of type MCQ whose data is represented by logical columns

TODO: other column needs to be handled}
}
\description{
This function takes an MCQ question represented by multiple logical columns and collapses it to a single column
}
