% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLabelFile_Source.R
\name{compareChars}
\alias{compareChars}
\title{Compares 2 strings by dividing the first grouping of consecutive shared characters by the total number of characters of the larger string}
\usage{
compareChars(str1, str2)
}
\arguments{
\item{str1:}{first string to be compared}

\item{str2:}{second string to be compared
TODO: maybe rework suppressWarnings call to only apply to length warning}
}
\description{
Compares 2 strings by dividing the first grouping of consecutive shared characters by the total number of characters of the larger string
}
