% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLabelFile_Source.R
\name{gatherMarbles}
\alias{gatherMarbles}
\title{Corrects qInds by examning if a single marble signifier has been split into
several questions due to dissimilar column titles}
\usage{
gatherMarbles(qInds, qClasses, questions)
}
\arguments{
\item{qInds:}{list in which each element contains start and end column indices
of each question}

\item{qClasses:}{List of classifications for each question

TODO: try to simplify algorithm.  Not a particuarly elegant solution even if it does work.}
}
\description{
Corrects qInds by examning if a single marble signifier has been split into
several questions due to dissimilar column titles
}
