% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLabelFile.R
\name{generateLabelFile}
\alias{generateLabelFile}
\title{Generates folder in working directory containing generated label file and text document with absolute path to input data}
\usage{
generateLabelFile(inputFile, threshold = 65)
}
\arguments{
\item{inputFile}{String containing path to .csv containing input data}

\item{threshold}{Integer between 0 and 100 which sets the threshold at which two column titles are considered part of the same question}
}
\description{
Generates folder in working directory containing generated label file and text document with absolute path to input data
}
