% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals_Source.R
\name{getMaxPlotId}
\alias{getMaxPlotId}
\title{helper function for savePlots() function. Obtains greatest ID number of .png images in a given plot folder}
\usage{
getMaxPlotId(type, path)
}
\arguments{
\item{type:}{A string idicating which type of question is represented in the given plots}

\item{path:}{Location of the output/visualizations/<plot type> folder}
}
\description{
helper function for savePlots() function. Obtains greatest ID number of .png images in a given plot folder
}
