% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals_Source.R
\name{getPlotDataCheckbox}
\alias{getPlotDataCheckbox}
\title{helper function for plotAll function family. contains code to transform data in preparation for violin plots
For checkbox MCQs.}
\usage{
getPlotDataCheckbox(.dyad, .factor)
}
\arguments{
\item{.dyad}{question object of type slider}

\item{.factor}{question object of type MCQ_Checkbox}
}
\description{
helper function for plotAll function family. contains code to transform data in preparation for violin plots
For checkbox MCQs.
}
