% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals_Source.R
\name{getPlotDataRadio}
\alias{getPlotDataRadio}
\title{helper function for plotAll function family. contains code to transform data in
preparation for violin plots for non-checkbox MCQs.}
\usage{
getPlotDataRadio(.dyad, .factor, plotInfo = NULL)
}
\arguments{
\item{.dyad:}{question object of type slider}

\item{.factor:}{question object of type MCQ}
}
\description{
helper function for plotAll function family. contains code to transform data in
preparation for violin plots for non-checkbox MCQs.
}
