% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLabelFile_Source.R
\name{getQIndices}
\alias{getQIndices}
\title{iterates through and creates list that holds the beginning and end index of each question}
\usage{
getQIndices(threshold, questions, answers)
}
\description{
iterates through and creates list that holds the beginning and end index of each question
}
