% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals_Source.R
\name{isCheckboxMcq}
\alias{isCheckboxMcq}
\title{Logical function which determines if a question is represented by logical columns or not}
\usage{
isCheckboxMcq(mcq)
}
\arguments{
\item{mcq:}{question object of type MCQ
essential logic is: if all columns (excluding NA and/or Other column if present) only contain the "1" character
the it's a checkbox}
}
\description{
Logical function which determines if a question is represented by logical columns or not
}
