% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLabelFile_Source.R
\name{isMcq}
\alias{isMcq}
\title{Determines if given set column names for a question describes a multiple choice question.  returns a logical
Helper to classifyQuestions()}
\usage{
isMcq(columns, answers)
}
\arguments{
\item{columns:}{list column names which describe a given question}
}
\description{
Determines if given set column names for a question describes a multiple choice question.  returns a logical
Helper to classifyQuestions()
}
