% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLabelFile_Source.R
\name{isSlider}
\alias{isSlider}
\title{Determines if given set column names for a question describes a slider signifier returns logical type
Helper to classifyQuestions()}
\usage{
isSlider(columns)
}
\arguments{
\item{columns:}{list column names which describe a given question}
}
\description{
Determines if given set column names for a question describes a slider signifier returns logical type
Helper to classifyQuestions()
}
