% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLabelFile_Source.R
\name{isTriangle}
\alias{isTriangle}
\title{Determines if given set column names for a question describes a triangle signifier returns logical type
Helper to classifyQuestions()}
\usage{
isTriangle(columns)
}
\arguments{
\item{columns:}{list column names which describe a given question}
}
\description{
Determines if given set column names for a question describes a triangle signifier returns logical type
Helper to classifyQuestions()
}
