% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgEnvDefinition.R
\docType{data}
\name{pkgGlobals}
\alias{pkgGlobals}
\title{pkgGlobals is an environment which exists in the namespace of the spryngr package.}
\format{
An object of class \code{environment} of length 1.
}
\usage{
pkgGlobals
}
\description{
pkgGlobals$outPath is a global variable in the package namespace which holds the
location of the output folder from which the question objects are being generated
}
\keyword{datasets}
