% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals.R
\name{plotAllDyads}
\alias{plotAllDyads}
\title{Generate a violin plot for each combination of Slider and MCQ Question object}
\usage{
plotAllDyads(dyads, factors)
}
\arguments{
\item{dyads}{A list of Question objects of type Slider}

\item{factors}{a list of Question Objects of type MCQ or MCQ_Checkbox}
}
\value{
a list of ggplot objects
}
\description{
Generate a violin plot for each combination of Slider and MCQ Question object
}
