% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals.R
\name{plotDyad}
\alias{plotDyad}
\title{Function which plots single dyad vs single factor which allows specification of labels and titles
Any labels left blank will be gathered from the generated labels document
Plot single dyad vs single factor with specification of labels and tites}
\usage{
plotDyad(dyad, factor, title = NULL, ylab = NULL, xlab = NULL)
}
\arguments{
\item{dyad:}{Question object of type Slider}

\item{factor:}{Question object of type MCQ or MCQ_Checkbox}

\item{title:}{String containing the title for the generated plot (optional)}

\item{ylab:}{String containing the Y label for the generated plot (optional)}

\item{xlab:}{Vector containing 2 Strings to label the x axis in the negative and positive direction respectively (optional)}
}
\value{
a ggplot object
}
\description{
Function which plots single dyad vs single factor which allows specification of labels and titles
Any labels left blank will be gathered from the generated labels document
Plot single dyad vs single factor with specification of labels and tites
}
