% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals.R
\name{plotMarble}
\alias{plotMarble}
\title{Plot single marble vs single factor with specification of sub-marble-index, axis, labels, and titles}
\usage{
plotMarble(
  marble,
  factor,
  varIndex,
  axis,
  title = NULL,
  ylab = NULL,
  xlab = NULL,
  save = FALSE
)
}
\arguments{
\item{save}{Logical indicating whether to save produced plots in output folder}

\item{marble:}{Question object of type Marble}

\item{factor:}{Question object of type MCQ or MCQ_Checkbox}

\item{varIndex:}{integer indicating the index of the Marble to be used}

\item{axis:}{integer equal to 1 or 2 dictating x or y axis respectively}

\item{title:}{String containing the title for the generated plot (optional)}

\item{ylab:}{String containing the Y label for the generated plot (optional)}

\item{xlab:}{Vector containing 2 Strings to label the x axis in the negative and positive direction respectively (optional)}
}
\value{
a ggplot object
}
\description{
Plot single marble vs single factor with specification of sub-marble-index, axis, labels, and titles
}
