% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals_Source.R
\name{plotQuestion}
\alias{plotQuestion}
\title{Helper function for plotAll function family. contains code to generate violin plots from
single dimension of continuous data over a set of factors}
\usage{
plotQuestion(data, .dyad, .factor, plotInfo = NULL)
}
\arguments{
\item{data:}{Data.frame containing data to generate plot}

\item{.dyad:}{question object of type slider}

\item{.factor:}{question object of type MCQ or MCQ_Checkbox}
}
\description{
Helper function for plotAll function family. contains code to generate violin plots from
single dimension of continuous data over a set of factors
}
