% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals.R
\name{plotTernary}
\alias{plotTernary}
\title{Plot single triangle vs single factor with specification of side, labels, and titles}
\usage{
plotTernary(
  ternary,
  factor,
  varIndex,
  title = NULL,
  ylab = NULL,
  xlab = NULL,
  save = FALSE
)
}
\arguments{
\item{save}{Logical indicating whether to save produced plots in output folder}

\item{ternary:}{Question object of type Triangle}

\item{factor:}{Question object of type MCQ or MCQ_Checkbox Question object}

\item{varIndex:}{int 1-3 indicating which side of the Triangle to use (1=left, 2=bottom, 3=right)}

\item{title:}{String containing the title for the generated plot (optional)}

\item{ylab:}{String containing the Y label for the generated plot (optional)}

\item{xlab:}{Vector containing 2 strings to label the x axis in the negative and positive direction respectively (optional)}
}
\value{
a ggplot object
}
\description{
Plot single triangle vs single factor with specification of side, labels, and titles
}
