% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals_Source.R
\name{prepMarbles}
\alias{prepMarbles}
\title{Helper function for plotAllMarbles.  Generates collection of
question objects for each column of each marble question}
\usage{
prepMarbles(marbles)
}
\arguments{
\item{marbles:}{list of marble objects}
}
\description{
Helper function for plotAllMarbles.  Generates collection of
question objects for each column of each marble question
}
