% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals_Source.R
\name{randString}
\alias{randString}
\title{helper function for writePlots() function.  Generates random selection of
consonants and numbers of length 5 for file names}
\usage{
randString()
}
\description{
helper function for writePlots() function.  Generates random selection of
consonants and numbers of length 5 for file names
}
