% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLabelFile_Source.R
\name{reshapeInds}
\alias{reshapeInds}
\title{takes a flat vector of indices and generates a list, each element containing
the beginning and ending indices for a set of question indices
helper for gatherMarbles()}
\usage{
reshapeInds(qLimits)
}
\arguments{
\item{qLimits:}{flat vector containing beginning and ending elements of marble signifiers}
}
\description{
takes a flat vector of indices and generates a list, each element containing
the beginning and ending indices for a set of question indices
helper for gatherMarbles()
}
