% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals_Source.R
\name{savePlots}
\alias{savePlots}
\title{helper function for plot function family. Function writes ggplots out to output folder}
\usage{
savePlots(plots, type, path)
}
\arguments{
\item{plots:}{Either a ggplot or list of ggplot objects}

\item{type:}{A string idicating which type of question is represented in the given plots}

\item{path:}{Location of the output/visualizations/<plot type> folder}
}
\description{
helper function for plot function family. Function writes ggplots out to output folder
}
