% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLabelFile_Source.R
\name{updateQuestionInds}
\alias{updateQuestionInds}
\title{Returns corrected qInd object
Helper for gatherMarbles()}
\usage{
updateQuestionInds(qInds, marbleInds, naInds)
}
\arguments{
\item{qInds:}{set of original question indices}

\item{marbleInds:}{list in which each element contains start and end indices
for marble signifiers which need to be gathered}

\item{naInds:}{Indices of NA columns for marble signifiers which have been incorrectly identified}
}
\description{
Returns corrected qInd object
Helper for gatherMarbles()
}
