% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLabelFile_Source.R
\name{writeLabelFile}
\alias{writeLabelFile}
\title{Iterate over each label, unlist group of labels, trim whitespace,
encase individual labels in single quotes,
Concattenate each vector of labels to single string,
write each of these strings out to xlsx file}
\usage{
writeLabelFile(outputName, qClassesDf, questions, answers)
}
\description{
Iterate over each label, unlist group of labels, trim whitespace,
encase individual labels in single quotes,
Concattenate each vector of labels to single string,
write each of these strings out to xlsx file
}
