% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals.R
\name{writePlots}
\alias{writePlots}
\title{Function which saves one or more ggplot objects}
\usage{
writePlots(plots, path = NULL, names = NULL, width = NA, height = NA)
}
\arguments{
\item{plots}{a ggplot object or list of ggplot objects}

\item{path}{a string containing a location to save the plots.  If left blank a location may be selected via a windows file explorer dialogue box}

\item{names}{a vector of strings the same length as the list of plots.  If left blank random names will be generated}
}
\value{

}
\description{
Function which saves one or more ggplot objects
}
