% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals.R
\name{generateQuestionObjects}
\alias{generateQuestionObjects}
\title{Generates a list of S4 Question objects for each question within label file}
\usage{
generateQuestionObjects(outputFolder)
}
\arguments{
\item{outputFolder}{String containing path to folder generated by call to generateLabelFile()}
}
\value{
A list containing four sublists titled
}
\description{
Generates a list of S4 Question objects for each question within label file
}
