% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVisuals.R
\name{plotAllMarbles}
\alias{plotAllMarbles}
\title{Generate a violin plot for each combination of Marble and MCQ Question object}
\usage{
plotAllMarbles(marbles, factors)
}
\arguments{
\item{marbles}{List of Question objects of type Marble}

\item{factors}{List of Question objects of type MCQ or MCQ_Checkbox}
}
\value{
a list of ggplot objects
}
\description{
Generate a violin plot for each combination of Marble and MCQ Question object
}
